updates mcp-server-giphy to follow new design
