+++ mcp-servers/mcp-server-giphy/.env.example
@@ -1,1 +1,5 @@
+# Optional for the service
+#LOG_LEVEL=DEBUG
+
+# Required for the service
 GIPHY_API_KEY=REPLACE_WITH_YOUR_GIPHY_API_KEY
