+++ mcp-servers/mcp-server-giphy/.vscode/launch.json
@@ -2,13 +2,13 @@
   "version": "0.2.0",
   "configurations": [
     {
-      "name": "mcp-servers: mcp-server-giphy",
       "type": "debugpy",
       "request": "launch",
-      "program": "${workspaceFolder}/server/main.py",
+      "name": "mcp-servers: mcp-server-giphy",
+      "cwd": "${workspaceFolder}",
+      "module": "mcp_server.start",
       "args": ["--transport", "sse", "--port", "6000"],
-      "console": "integratedTerminal",
-      "consoleName": "mcp-server-giphy"
+      "consoleTitle": "mcp-server-giphy"
       // "justMyCode": false // Set to false to debug external libraries
     }
   ]
