+++ mcp-servers/mcp-server-giphy/mcp_server/__init__.py
@@ -1,0 +1,6 @@
+from dotenv import load_dotenv
+from . import config
+
+# Load environment variables from .env into the settings object.
+load_dotenv()
+settings = config.Settings()
