+++ mcp-servers/mcp-server-giphy/mcp_server/sampling.py
