+++ mcp-servers/mcp-server-giphy/mcp_server/start.py
@@ -1,0 +1,31 @@
+# Main entry point for the MCP Server
+
+import argparse
+
+from .server import create_mcp_server
+
+def main() -> None:
+    # Command-line arguments for transport and port
+    parse_args = argparse.ArgumentParser(description=f"Start the MCP server.")
+    parse_args.add_argument(
+        "--transport",
+        default="stdio",
+        choices=["stdio", "sse"],
+        help="Transport protocol to use ('stdio' or 'sse'). Default is 'stdio'.",
+    )
+    parse_args.add_argument(
+        "--port",
+        type=int,
+        default=8000,
+        help="Port to use for SSE (default is 8000)."
+    )
+    args = parse_args.parse_args()
+
+    mcp = create_mcp_server()
+    if args.transport == "sse":
+        mcp.settings.port = args.port
+
+    mcp.run(transport=args.transport)
+
+if __name__ == "__main__":
+    main()
