--- mcp-servers/mcp-server-giphy/server/__init__.py
@@ -1,1 +1,0 @@
-# Init file for MCP GIPHY Server
