+++ mcp-servers/mcp-server-open-deep-research/.vscode/settings.json
@@ -63,6 +63,7 @@
     "huggingface",
     "KHTML",
     "pubchempy",
+    "SERP",
     "serpapi",
     "smolagents",
     "toplevel"
