+++ semantic-workbench.code-workspace
@@ -156,12 +156,16 @@
       "path": "mcp-servers/mcp-server-giphy"
     },
     {
-      "name": "mcp-servers:mcp-server-research",
-      "path": "mcp-servers/mcp-server-research"
+      "name": "mcp-servers:mcp-server-open-deep-research",
+      "path": "mcp-servers/mcp-server-open-deep-research"
     },
     {
       "name": "mcp-servers:mcp-server-vscode",
       "path": "mcp-servers/mcp-server-vscode"
+    },
+    {
+      "name": "mcp-servers:mcp-server-template",
+      "path": "mcp-servers/mcp-server-template"
     },
     {
       "name": "tools",
